+++ ngx-fudis/projects/ngx-fudis/src/lib/components/dialog/dialog.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { MatDialogModule } from '@angular/material/dialog';
 
-import { FudisDialogService } from './dialog.service';
+import { FudisDialogService } from '../../services/dialog/dialog.service';
 
 describe('FudisDialogServiceService', () => {
 	let service: FudisDialogService;
